fix init flag

+++ models/xlnet/__init__.py
@@ -133,9 +133,9 @@ class Embeddings(object):
         xlnet_model = xlnet.XLNetModel(
             xlnet_config=cls.xlnet_config,
             run_config=cls.run_config,
-            input_ids=np.array(input_ids),
-            seg_ids=np.array(segment_ids),
-            input_mask=np.array(input_masks, dtype=np.int64))
+            input_ids=np.array(input_ids, dtype=np.int32),
+            seg_ids=np.array(segment_ids, dtype=np.int32),
+            input_mask=np.array(input_masks, dtype=np.int32))
 
         # Get a sequence output
         sequence_output = xlnet_model.get_sequence_output()
